refactor(bench): extract the fold fixture and sampler into bench/lib
`bench/fold-cost.ts` ends in an unguarded `main().then(process.exit)`, so
importing any symbol from it runs the whole 12-cell bench and then exits
the host process. That makes its fixture builder and CPU/heap sampler
unreusable: the fold-ceiling probe needs them, and a test that imports
them would have its vitest worker killed mid-suite.

Move `buildFixture`, `measureOneFold`, `median`, and the constants they
close over into a side-effect-free `bench/lib/fold-fixture.ts`, and have
`fold-cost.ts` import them. This is the split `bench/amortized-write-cost.ts`
and `bench/write-amp-stress.ts` already use over
`bench/lib/write-cost-measure.ts` — runner scripts own the grid and the
output format, the library owns the measurement.

Pure move: no signature, seed, or measurement-definition change. Verified
behaviour-free by running the bench before and after and diffing the
deterministic columns (axis, label, rows, bytes_per_doc, snapshot_bytes)
across all 12 cells — identical, and still matching
`docs/spec/attachments/fold-cost-baseline.json` cell for cell.
