+++ bench/fold-cost.ts
@@ -28,7 +28,13 @@
  * Storage is in-memory so I/O is ~free and the measured cost is the
  * CPU/allocation of the rebuild itself.
  *
- * MEASUREMENT METHOD
+ * The fixture builder and the per-fold CPU/heap measurement live in
+ * `bench/lib/fold-fixture.ts` so the fold-ceiling probe can reuse them
+ * verbatim — same seed, same tail, same measurement definitions — and so
+ * they are importable without running this bench. This file owns the
+ * grid, the iteration counts, and the output format.
+ *
+ * MEASUREMENT METHOD (defined in `bench/lib/fold-fixture.ts`)
  *   - **CPU**: `process.cpuUsage()` deltas (user + system µs) bracketing
  *     the fold, reported in ms. NOT wall-clock — folds are CPU-bound on
  *     Workers and wall would include I/O (which MemoryStorage makes ~free
@@ -76,51 +82,23 @@
  * portable signal.
  */
 
-/* eslint-disable no-underscore-dangle -- `_id` is the locked primary-key
-   field on document + snapshot shapes (see `@baerly/protocol`'s
-   `Collection<T>` / the snapshot body). */
-
 import { mkdir, writeFile } from "node:fs/promises";
 import path from "node:path";
+import { MAINTENANCE_MAX_FOLD_ROWS } from "@baerly/protocol";
 import {
-  type CurrentJson,
-  type DocumentData,
-  type LogEntry,
-  MAINTENANCE_MAX_FOLD_ROWS,
-  MemoryStorage,
-  countKey,
-  encodeJsonBytes,
-  snapshotHash,
-  timestamp,
-} from "@baerly/protocol";
-import { type SnapshotBody, encodeSnapshotBody, snapshotKey } from "@baerly/server";
-import { compact } from "@baerly/server/maintenance";
-import type { InternalCompactOptions } from "@baerly/server/_internal/testing";
+  HEAP_SAMPLE_INTERVAL_MS,
+  SEED,
+  TAIL_ENTRIES,
+  buildFixture,
+  measureOneFold,
+  median,
+} from "./lib/fold-fixture.ts";
 
 // ── Bench config. Pinned constants — tweak in the source if needed. ──
-
-const SEED = 0xf01d_c057; // "fold cost"; reproduction handle.
-const SESSION = "fold000"; // 7 chars; the bench doesn't validate sessions.
-const COLLECTION = "notes";
-const CURRENT_JSON_KEY = `app/x/tenant/t/manifests/${COLLECTION}/current.json`;
-const COLLECTION_PREFIX = `app/x/tenant/t/manifests/${COLLECTION}`;
 
 /** Warmup folds (discarded) then measured folds, per grid point. */
 const WARMUP_ITERS = 5;
 const MEASURE_ITERS = 11;
-
-/** Heap-sampling interval during a fold (ms). */
-const HEAP_SAMPLE_INTERVAL_MS = 1;
-
-/**
- * Tail length folded per measured rebuild. Small + fixed so the fold's
- * cost is dominated by the snapshot rebuild (load old + serialize new +
- * hash), not by the tail walk — which mirrors the production shape (the
- * tail is SLICED, the snapshot rebuild is the unsliceable cost). The
- * tail updates existing docs so the new snapshot stays the same size /
- * row count as the old (a steady-state fold).
- */
-const TAIL_ENTRIES = 100;
 
 /**
  * BYTES axis — fixed bytes/doc, row count chosen to bracket the
@@ -154,177 +132,6 @@ const ROWS_AXIS_ROW_COUNTS: readonly number[] = [
   8192,
   16384,
 ];
-
-/**
- * Mulberry32 PRNG — small, seedable, no deps. Deterministic input so
- * two runs on the same machine produce comparable snapshot shapes.
- */
-const mulberry32 = (seed: number): (() => number) => {
-  let a = seed >>> 0;
-  return () => {
-    a = (a + 0x6d2b79f5) >>> 0;
-    let t = a;
-    t = Math.imul(t ^ (t >>> 15), t | 1);
-    t ^= t + Math.imul(t ^ (t >>> 7), t | 61);
-    return ((t ^ (t >>> 14)) >>> 0) / 4294967296;
-  };
-};
-
-/**
- * Build a document whose canonical JSON byteLength is close to
- * `targetBytes`. The doc carries a few typed fields plus a `pad` string
- * sized to hit the target; representative of a real notes-shaped record
- * (string body + scalars) rather than one giant blob.
- */
-const makeDoc = (id: string, targetBytes: number, rng: () => number): DocumentData => {
-  // Fixed scaffold the encoder always emits; the pad absorbs the rest.
-  const scaffold: DocumentData = {
-    _id: id,
-    title: `note ${id}`,
-    n: Math.floor(rng() * 1_000_000),
-    done: rng() < 0.5,
-    pad: "",
-  };
-  const scaffoldBytes = encodeJsonBytes(scaffold).byteLength;
-  const padLen = Math.max(0, targetBytes - scaffoldBytes);
-  // Printable ASCII so each char is one JSON byte (no escaping / multi-
-  // byte surprises that would throw off the target).
-  return { ...scaffold, pad: "x".repeat(padLen) };
-};
-
-interface FoldFixture {
-  readonly storage: MemoryStorage;
-  /** Actual canonical snapshot-body byteLength (measured, not target). */
-  readonly snapshotBytes: number;
-  readonly rows: number;
-}
-
-/**
- * Seed a fresh `MemoryStorage` with a `current.json` + a prior snapshot
- * of (rows × bytesPerDoc) + a `TAIL_ENTRIES`-long log tail that updates
- * existing docs. After this returns, a single `compact()` call folds the
- * whole tail into a rebuilt snapshot of the same shape.
- */
-const buildFixture = async (rows: number, bytesPerDoc: number): Promise<FoldFixture> => {
-  const storage = new MemoryStorage();
-  const rng = mulberry32(SEED ^ (rows * 0x9e37_79b1) ^ bytesPerDoc);
-
-  // 1. Prior snapshot: `rows` docs, sorted by _id (compactor invariant).
-  const docs: Array<{ _id: string; body: DocumentData }> = [];
-  for (let i = 0; i < rows; i++) {
-    const id = `doc-${i.toString().padStart(8, "0")}`;
-    docs.push({ _id: id, body: makeDoc(id, bytesPerDoc, rng) });
-  }
-  docs.sort((a, b) => byIdAsc(a._id, b._id));
-  const snapBody: SnapshotBody = {
-    schema_version: 1,
-    min_seq: 0,
-    max_seq: rows, // prior snapshot covered [0, rows)
-    collection: COLLECTION,
-    docs,
-  };
-  const snapBytes = encodeSnapshotBody(snapBody);
-  const sha = await snapshotHash(snapBytes);
-  const snapKey = snapshotKey(COLLECTION_PREFIX, 0, rows, sha);
-  await storage.put(snapKey, snapBytes, { contentType: "application/json" });
-
-  // 2. Log tail: TAIL_ENTRIES updates to existing docs at seqs
-  //    [rows, rows + TAIL_ENTRIES). U with a full post-image keeps the
-  //    rebuilt snapshot the same row count (steady-state fold).
-  for (let t = 0; t < TAIL_ENTRIES; t++) {
-    const seq = rows + t;
-    const targetIdx = Math.floor(rng() * rows);
-    const id = `doc-${targetIdx.toString().padStart(8, "0")}`;
-    const entry: LogEntry = {
-      lsn: `${timestamp(1_700_000_000_000 + seq)}_${SESSION}_${countKey(seq)}`,
-      commit_ts: new Date(1_700_000_000_000 + seq).toISOString(),
-      op: "U",
-      collection: COLLECTION,
-      doc_id: id,
-      after: makeDoc(id, bytesPerDoc, rng),
-      session: SESSION,
-      seq,
-    };
-    const entryBytes = encodeJsonBytes(entry);
-    await storage.put(`${COLLECTION_PREFIX}/log/${seq}.json`, entryBytes, {
-      contentType: "application/json",
-    });
-  }
-
-  // 3. current.json pointing at the prior snapshot with the tail live.
-  const current: CurrentJson = {
-    schema_version: 3,
-    snapshot: snapKey,
-    tail_hint: rows + TAIL_ENTRIES,
-    log_seq_start: rows,
-    writer_fence: { epoch: 1, owner: "fold-cost-bench", claimed_at: "" },
-    snapshot_bytes: snapBytes.byteLength,
-    snapshot_rows: rows,
-  };
-  await storage.put(CURRENT_JSON_KEY, encodeJsonBytes(current), {
-    ifNoneMatch: "*",
-    contentType: "application/json",
-  });
-
-  return { storage, snapshotBytes: snapBytes.byteLength, rows };
-};
-
-/** A single measured fold over an already-built fixture. */
-const measureOneFold = async (
-  storage: MemoryStorage,
-): Promise<{ cpuMs: number; peakBytes: number }> => {
-  const heapStart = process.memoryUsage().heapUsed;
-  let peak = heapStart;
-  const sampler = setInterval(() => {
-    const h = process.memoryUsage().heapUsed;
-    if (h > peak) {
-      peak = h;
-    }
-  }, HEAP_SAMPLE_INTERVAL_MS);
-  // `unref` so a stray timer can never keep the process alive.
-  sampler.unref();
-
-  const cpu0 = process.cpuUsage();
-  // Whole tail in one pass (the unsliceable rebuild); no ceiling (we are
-  // measuring the fold, not the defer path). `maxEntriesPerRun` rides on
-  // the internal options object.
-  const opts: InternalCompactOptions = {
-    minEntriesToCompact: 1,
-    maxEntriesPerRun: Number.MAX_SAFE_INTEGER,
-  };
-  const res = await compact({ storage, currentJsonKey: CURRENT_JSON_KEY }, opts);
-  const cpu1 = process.cpuUsage(cpu0);
-  clearInterval(sampler);
-
-  if (!res.written) {
-    throw new Error(`fold-cost: expected a written fold, got skippedReason=${res.skippedReason}`);
-  }
-  // One last sample in case the fold finished between ticks.
-  const heapEnd = process.memoryUsage().heapUsed;
-  if (heapEnd > peak) {
-    peak = heapEnd;
-  }
-  const cpuMs = (cpu1.user + cpu1.system) / 1000; // µs → ms
-  const peakBytes = Math.max(0, peak - heapStart);
-  return { cpuMs, peakBytes };
-};
-
-const median = (xs: readonly number[]): number => {
-  const s = [...xs].toSorted((a, b) => a - b);
-  const mid = Math.floor(s.length / 2);
-  return s.length % 2 === 0 ? (s[mid - 1]! + s[mid]!) / 2 : s[mid]!;
-};
-
-/** Lexicographic `_id` comparator — same ordering the compactor uses. */
-const byIdAsc = (a: string, b: string): number => {
-  if (a < b) {
-    return -1;
-  }
-  if (a > b) {
-    return 1;
-  }
-  return 0;
-};
 
 /** Bytes → KB / MB for the summary table. */
 const kb = (b: number): string => (b / 1024).toFixed(0);
